% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{fw_create}
\alias{fw_create}
\title{Create Foodweb model object}
\usage{
fw_create(A, B, R, U = NULL, sdB = NULL)
}
\arguments{
\item{A}{interaction matrix. Either a signed matrix (0/-1/1) or a matrix
with coefficients. In both cases the matrix is currently treated as a signed
one.}

\item{B}{biomass vector.}

\item{R}{reproduction/mortality vector.}

\item{U}{unknow matrix. A two-columns matrix that includes unknowm column.}

\item{sdB}{vector of standard deviation biomass (see \code{\link[limSolve:xsample]{limSolve::xsample()}}).}
}
\value{
An object of class \code{class} which is a named list of 6 elements:
\itemize{
\item A
\item B
\item R
\item U
\item SdB
\item model: the underlying model
}
}
\description{
Create Foodweb model object
}
