% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_case.R
\name{fw_example_2species}
\alias{fw_example_2species}
\alias{fw_example_5species_2chains}
\title{Study case}
\usage{
fw_example_2species()

fw_example_5species_2chains(add_sdB = FALSE)
}
\arguments{
\item{add_sdB}{a logical. Should biomass variation be added?}
}
\description{
Helper function that includes study cases.
}
\examples{
fw_example_2species()
}
