% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_infer.R
\name{fw_infer}
\alias{fw_infer}
\title{Estimate interaction strengths}
\usage{
fw_infer(A, R, B, eff_max = 1)
}
\arguments{
\item{A}{interaction matrix}

\item{R}{reproduction/mortality.}

\item{B}{biomass matrix.}

\item{eff_max}{maximum transfer efficiency.}
}
\description{
Function to estimate interaction strengths using LIM.
}
\details{
LIM based on generalized linear Lotka-Volterra model with the following
interaction matrix:
\deqn{\frac{1/X}{X'} = A * X + R }{1/X X' = A * X + R}
See \code{\link[limSolve:xsample]{limSolve::xsample()}} for the meaning of matrices E, F, G and H.
}
\references{
\itemize{
\item Gellner G, McCann K, Hastings A. 2023. Stable diverse food webs become more common when interactions are more biologically constrained. Proceedings of the National Academy of Sciences 120:e2212061120. DOI: 10.1073/pnas.2212061120.
}
}
