% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_model.R
\name{fw_model}
\alias{fw_model}
\alias{fw_get_model}
\title{Create a foodweb model object}
\usage{
fw_model(A, B, R)

fw_get_model(model = c("lv_fr1"))
}
\arguments{
\item{A}{interaction matrix. A square matrix describing interactions of the
system (see details).}

\item{B}{biomass vector.}

\item{R}{reproduction/mortality vector.}

\item{model}{model to be used.}
}
\value{
An object of class \code{fw_problem} which is a named list of 4 elements:
\itemize{
\item A: interaction matrix
\item B: biomass vector
\item R: mortality rate vector
\item model: the underlying model
}
}
\description{
Create an object of class \code{fw_model}.
}
\details{
So far there is only one model supported, a general Lotka-Volterra with
a linear response.
}
\section{Functions}{
\itemize{
\item \code{fw_get_model()}: get model

}}
\examples{
A <- rbind(c(-1, -1), c(1, 0))
R <- c(0.1, -0.05)
B <- c(0.5, 0.25)
fw_model(A, B, R)
}
