% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sfw.R
\name{fw_niche_model}
\alias{fw_niche_model}
\title{Helper to generate structured model foodwebs}
\usage{
fw_niche_model(
  nsp,
  connectance,
  connect_all = FALSE,
  unbias = FALSE,
  niche = NULL
)
}
\arguments{
\item{nsp}{An integer giving the number of species considered.}

\item{connectance}{A real positive between 0 and .5 indicating the
connectance of the network to be generated.}

\item{connect_all}{Logical. If \code{TRUE}, then all species in the network have a
least one prey (but the niche with the lowest niche value).}

\item{unbias}{Logical. If \code{TRUE}, then the first species may not be a basal
species.}

\item{niche}{A vector real positive between 0 and 1 standing for the niche
axis. Default is set to \code{NULL}, in such case the niche axis is automatically
generated.}
}
\value{
A logical matrix describing pairwise interactions. A given line
describes the diet of a given species while a column describes the set of
predator associated to a particular species.
}
\description{
This function is an implementation of the niche model (Williams & Martinez,
2000) which generates food webs based on three elements:
\itemize{
\item a number of species,
\item a niche axis and
\item a value of connectance.
}
}
\details{
Three remarks. First, according to Williams and Martinez (2000),
the species with the lowest niche value is considered as a basal species and
therefore cannot feed upon another species. This introduces a slight bias
(\emph{e.g} the expected connectance is lower than the expected values
\code{connectance}). Second, forcing all the species to be connected introduces
another bias (on connectance values) as they tends to be more connected than
expected. Third, if with custom niche axes, the expected connectance may not
be the one expected (\code{connectance}) as niche values distribution will likely
differs from the uniform distribution used in Williams and Martinez (2000).
}
\examples{
fw_niche_model(10, 0.2)
}
\references{
Williams, R.J. and Martinez, N.D. (2000) Simple rules yield complex food
webs. \emph{Nature}, 404:180–183.
}
