% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{fw_ode_plot}
\alias{fw_ode_plot}
\alias{fw_range_plot}
\title{Simple ODE plot}
\usage{
fw_ode_plot(x, init, times)

fw_range_plot(y, include_stability = TRUE)
}
\arguments{
\item{x}{an object of class \code{fw_model}.}

\item{init}{initial values (see \code{\link[deSolve:ode]{deSolve::ode()}}).}

\item{times}{time steps (see \code{\link[deSolve:ode]{deSolve::ode()}}).}

\item{y}{an object of class \code{fw_predict}.}

\item{include_stability}{a logical. Should the distribution of eigen values
be included?}
}
\description{
Simple ODE plot
}
\examples{
fw_ode_plot(fw_example_2species(), rep(0.3, 2), seq(1, 500, 0.1))
}
