% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_problem.R
\name{fw_problem}
\alias{fw_problem}
\alias{fw_as_problem}
\title{Create Foodweb model object}
\usage{
fw_problem(A, B, R, U = NULL, sdB = NULL)

fw_as_problem(x, ...)
}
\arguments{
\item{A}{interaction matrix. A square matrix describing interactions of the
system (see details).}

\item{B}{biomass vector.}

\item{R}{reproduction/mortality vector.}

\item{U}{unknow interactions. A three-columns data frame that includes the
position of unknown/known interactions. If \code{NULL}, U is derived from A.}

\item{sdB}{vector of standard deviation biomass (see \code{\link[limSolve:xsample]{limSolve::xsample()}}).}

\item{x}{an R object.}

\item{...}{extra arguments (ignored so far, might be use to custom xsample
at some point).}
}
\value{
An object of class \code{fw_problem} which is a named list of 7 elements:
\itemize{
\item A: interaction matrix
\item B: biomass vector
\item R: mortality rate vector
\item U: unknown interaction(s)
\item S: square matrix representing the signed structure (only 1/-1 and 0)
\item SdB: biomass variation
\item model: the underlying model
}
}
\description{
Create an object of class \code{fw_problem} that can be used for interaction
inference.
}
\details{
In most cases, only \code{A} needs to be defined and should include:
\itemize{
\item 1 when interactions are present and positve,
\item -1 when interactions are present and negative,
\item 0 otherwise.
\code{U} will be generated according to this these values. \code{A} may however
includes known interactions, in such case, \code{U} needs to be defined and
must includes all known interactions (unknown interactions will be
identified using \code{A}).
TODO: S may not be required if U has signed info, or U may not be required if
S points to only unknown info. I think ultimately using an Edge list + Node
info is the way to go.
}
}
\section{Functions}{
\itemize{
\item \code{fw_as_problem()}: Coerce an object to an object of class \code{fw_problem}.

}}
\examples{
A <- rbind(c(-1, -1), c(1, 0))
R <- c(0.1, -0.05)
B <- c(0.5, 0.25)
fw_problem(A, B, R)

}
