% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{fw_infer}
\alias{fw_infer}
\alias{fw_predict_A}
\alias{fw_predict_B}
\title{Estimate interaction strengths}
\usage{
fw_infer(x, eff_max = 1, ...)

fw_predict_A(y, index = 1)

fw_predict_B(y, index = 1)
}
\arguments{
\item{x}{an object of class \code{fw_problem}.}

\item{eff_max}{max efficiency.}

\item{...}{further arguments passed to \code{\link[limSolve:xsample]{limSolve::xsample()}}).}

\item{y}{an object of class \code{fw_predicted}}

\item{index}{index of the prediction to be used.}
}
\value{
Return a list of two elements:
\itemize{
\item \code{prediction}: a data frame with one column per interaction strength
estimated. The number of row is given by the number of interaction sets
estimated by \code{xsample()} (see parameter \code{iter} in \code{xsample()}).
}
}
\description{
Function to estimate interaction strengths using LIM.
}
\details{
LIM based on generalized linear Lotka-Volterra model with the following
interaction matrix:
\deqn{\frac{1/X}{X'} = A * X + R }{1/X X' = A * X + R}
See \code{\link[limSolve:xsample]{limSolve::xsample()}} for the meaning of matrices E, F, G and H.
}
\section{Functions}{
\itemize{
\item \code{fw_predict_A()}: returns predicted A.

\item \code{fw_predict_B()}: returns predicted B.

}}
\examples{
A <- rbind(c(-1, -1), c(1, 0))
R <- c(0.1, -0.05)
B <- c(0.5, 0.25)
res <- fw_problem(A, B, R) |> fw_infer()

}
\references{
\itemize{
\item Gellner G, McCann K, Hastings A. 2023. Stable diverse food webs become
more common when interactions are more biologically constrained. Proceedings
of the National Academy of Sciences 120:e2212061120. DOI: 10.1073/pnas.2212061120.
}
}
