% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_infer.R
\name{fw_infer}
\alias{fw_infer}
\title{Estimate interaction strengths}
\usage{
fw_infer(A, R, B, eff_max = 1, sdB = NULL, ...)
}
\arguments{
\item{A}{interaction matrix. Either a signed matrix or a matrix with signed
coefficient. In both cases the matrix is currently treated as a signed one.}

\item{R}{reproduction/mortality vector.}

\item{B}{biomass vector.}

\item{eff_max}{maximum transfer efficiency.}

\item{sdB}{vector of standard deviation biomass (see \code{\link[limSolve:xsample]{limSolve::xsample()}}).}

\item{...}{further arguments passed to \code{\link[limSolve:xsample]{limSolve::xsample()}}).}
}
\value{
Return a data frame with one column per interaction strength estimated in
A (i.e. one per 1 or -1 in A). The number of row is given by the number of
interaction estimated by \code{xsample()} (see parameter \code{iter} in \code{xsample()}).
}
\description{
Function to estimate interaction strengths using LIM.
}
\details{
LIM based on generalized linear Lotka-Volterra model with the following
interaction matrix:
\deqn{\frac{1/X}{X'} = A * X + R }{1/X X' = A * X + R}
See \code{\link[limSolve:xsample]{limSolve::xsample()}} for the meaning of matrices E, F, G and H.
}
\references{
\itemize{
\item Gellner G, McCann K, Hastings A. 2023. Stable diverse food webs become more common when interactions are more biologically constrained. Proceedings of the National Academy of Sciences 120:e2212061120. DOI: 10.1073/pnas.2212061120.
}
}
