% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{fw_problem}
\alias{fw_problem}
\alias{fw_as_problem}
\title{Create Foodweb model object}
\usage{
fw_problem(A, B, R, U = NULL, sdB = NULL)

fw_as_problem(x, ...)
}
\arguments{
\item{A}{interaction matrix. A square matrix describing interactions of the
system (see details). If \code{NULL}, \code{A} is derived from \code{U}.}

\item{B}{biomass vector.}

\item{R}{reproduction/mortality vector.}

\item{U}{unknown interactions. A data frame that includes the
position of unknown/known interactions. If \code{NULL}, \code{U} is derived from \code{A}.}

\item{sdB}{vector of standard deviation biomass (see \code{\link[limSolve:xsample]{limSolve::xsample()}}).}

\item{x}{an R object.}

\item{...}{extra arguments (ignored so far, might be use to custom xsample
at some point).}
}
\value{
An object of class \code{fw_problem} which is a named list of 7 elements:
\itemize{
\item A: interaction matrix
\item B: biomass vector
\item R: mortality rate vector
\item U: unknown interaction(s)
\item SdB: biomass variation
\item model: the underlying model
}
}
\description{
Create an object of class \code{fw_problem} that can be used for interaction
inference.
}
\details{
In most cases, only \code{A} needs to be defined and should include:
\itemize{
\item 1 when interactions are present and positive,
\item -1 when interactions are present and negative,
\item 0 otherwise.
\code{U} will be generated according to these values. \code{A} may however
includes known interactions, in such case, \code{U} needs to be defined and
must includes all known interactions (unknown interactions will be
identified using \code{A}).
It is also possible to only declare \code{U}, in this case the dimension of A and
its coefficients will be derived from \code{U}. \code{U} has a strict format, namely
a data frame with 5 columns:
\item \code{name}: interaction name, when \code{U} is created from \code{A} name is \code{a_i_j}
where \code{i} is the row number and \code{j} the column number,
\item \code{row}: row number,
\item \code{col}: column number,
\item \code{unknown}: a logical vector that identifies unknown and known interactions,
\item \code{value}: interaction value, only the sign matter for unknown interactions,
The order does not matter as the columns are ordered during the validation
step.
}
}
\section{Functions}{
\itemize{
\item \code{fw_as_problem()}: Coerce an object to an object of class \code{fw_problem}.

}}
\examples{
A <- rbind(c(-1, -1), c(1, 0))
R <- c(0.1, -0.05)
B <- c(0.5, 0.25)
fw_problem(A, B, R)

}
