% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{webFromNicheModel}
\alias{webFromNicheModel}
\title{Core implementation of the niche model.}
\usage{
webFromNicheModel(nsp, connec, niche, connect_all = FALSE, unbias = FALSE)
}
\description{
Core implementation of the niche model.
}
